Add FRP interface to Brick
This adds support for FRP-style UIs using Brick datatypes.
Unfortunately, viewports are not (yet?) supported.

The entry point for reactive applications is

    reactiveMain :: ReactiveBuilder a -> IO a

..where the "ReactiveBuilder" describes how to build a ReactiveApp from
Vty-based reactive Events

ReactiveApp contains a Behavior and an Event:

    bView :: Behavior RenderView
    eExit :: Event a

RenderView is a snapshot of the current widget layers, cursor location,
and attribute map, making `bView` analogous to Brick's `App` data type.

When an event is pulled from eExit, the UI will be destroyed and the
`reactiveMain` IO action will return the event's value.